Refactor was a breeze with Visual Studio Code and extensions
