+++ src/prefabs.rs
@@ -1,0 +1,6 @@
+pub mod background;
+pub mod bird;
+pub mod button;
+pub mod ground;
+pub mod pipes;
+pub mod scoreboard;
