+++ src/scenes.rs
@@ -1,0 +1,2 @@
+pub mod title;
+pub mod game;
