+++ src/systems/scenemanagement.rs
@@ -1,0 +1,65 @@
+use tetra::graphics::ScreenScaling;
+use tetra::window;
+use tetra::{Context, State};
+use tetra::graphics::{self};
+
+use crate::scenes::title::{TitleScene};
+
+// === Scene Management ===
+
+pub trait Scene {
+    fn update(&mut self, ctx: &mut Context) -> tetra::Result<Transition>;
+    fn draw(&mut self, ctx: &mut Context, dt: f64);
+}
+
+pub enum Transition {
+    None,
+    Push(Box<dyn Scene>),
+    Pop,
+}
+
+// Boxing/dynamic dispatch could be avoided here by defining an enum for all
+// of your scenes, but that adds a bit of extra boilerplate - your choice!
+
+pub struct SceneManager {
+    scenes: Vec<Box<dyn Scene>>,
+}
+
+impl SceneManager {
+    pub fn new(ctx: &mut Context) -> tetra::Result<SceneManager> {
+        let initial_scene = TitleScene::new(ctx)?;
+        graphics::set_scaling(ctx, ScreenScaling::ShowAllPixelPerfect);
+        window::show_mouse(ctx);
+        Ok(SceneManager {
+            scenes: vec![Box::new(initial_scene)],
+        })
+    }
+}
+
+impl State for SceneManager {
+    fn update(&mut self, ctx: &mut Context) -> tetra::Result {
+        match self.scenes.last_mut() {
+            Some(active_scene) => match active_scene.update(ctx)? {
+                Transition::None => {}
+                Transition::Push(s) => {
+                    self.scenes.push(s);
+                }
+                Transition::Pop => {
+                    self.scenes.pop();
+                }
+            },
+            None => window::quit(ctx),
+        }
+
+        Ok(())
+    }
+
+    fn draw(&mut self, ctx: &mut Context, dt: f64) -> tetra::Result {
+        match self.scenes.last_mut() {
+            Some(active_scene) => active_scene.draw(ctx, dt),
+            None => window::quit(ctx),
+        }
+
+        Ok(())
+    }
+}
